+++ source-code/cljs/x/app_elements/combo_box.cljs
@@ -276,7 +276,6 @@
   ;   :default-value (*)(opt)
   ;   :disabled? (boolean)(opt)
   ;    Default: false
-  ;   :disabler (subscription-vector)(opt)
   ;   :emptiable? (boolean)(opt)
   ;    Default: true
   ;   :extendable? (boolean)(opt)