+++ source-code/cljs/x/app_elements/counter.cljs
@@ -145,7 +145,6 @@
   ;   :default-value (integer)(constant)(opt)
   ;   :disabled? (boolean)(opt)
   ;    Default: false
-  ;   :disabler (subscription-vector)(opt)
   ;   :form-id (keyword)(opt)
   ;   :helper (metamorphic-content)(opt)
   ;   :indent (keyword)(opt)