+++ source-code/cljs/x/app_elements/icon_button.cljs
@@ -5,8 +5,8 @@
 ; Author: bithandshake
 ; Created: 2020.10.16
 ; Description:
-; Version: v1.0.6
-; Compatibility: x4.5.5
+; Version: v1.1.0
+; Compatibility: x4.5.7
 
 
 
@@ -104,6 +104,20 @@
                       [icon-button-body            button-id button-props]
                       [engine/element-info-tooltip button-id button-props]])
 
+(defn- stated-element
+  ; WARNING! NON-PUBLIC! DO NOT USE!
+  ;
+  ; @param (keyword) button-id
+  ; @param (map) button-props
+  ;
+  ; @return (component)
+  [button-id button-props]
+  [engine/stated-element button-id
+                         {:render-f      #'icon-button
+                          :element-props button-props
+                          :destructor    [:elements/destruct-clickable! button-id]
+                          :initializer   [:elements/init-clickable!     button-id]}])
+
 (defn element
   ; @param (keyword)(opt) button-id
   ; @param (map) button-props
@@ -120,7 +134,6 @@
   ;    Only w/ {:variant :transparent}
   ;   :disabled? (boolean)(opt)
   ;    Default: false
-  ;   :disabler (subscription-vector)(opt)
   ;   :icon (keyword)
   ;   :icon-family (keyword)(opt)
   ;    :material-icons-filled, :material-icons-outlined
@@ -158,11 +171,8 @@
   ([button-props]
    [element (a/id) button-props])
 
-  ([button-id button-props]
+  ([button-id {:keys [keypress] :as button-props}]
    (let [button-props (engine/apply-preset    BUTTON-PROPS-PRESETS button-props)
          button-props (button-props-prototype button-props)]
-        [engine/stated-element button-id
-                               {:render-f      #'icon-button
-                                :element-props button-props
-                                :destructor    [:elements/destruct-clickable! button-id]
-                                :initializer   [:elements/init-clickable!     button-id]}])))
+        (if keypress [stated-element button-id button-props]
+                     [button         button-id button-props]))))