+++ source-code/cljs/x/app_elements/password_field.cljs
@@ -81,7 +81,6 @@
   ;   :default-value (string)(constant)(opt)
   ;   :disabled? (boolean)(opt)
   ;    Default: false
-  ;   :disabler (subscription-vector)(opt)
   ;   :form-id (keyword)(opt)
   ;   :indent (keyword)(opt)
   ;    :left, :right, :both, :none