+++ source-code/cljs/x/app_elements/search_field.cljs
@@ -59,7 +59,6 @@
   ;   :default-value (string)(constant)(opt)
   ;   :disabled? (boolean)(opt)
   ;    Default: false
-  ;   :disabler (subscription-vector)(opt)
   ;   :form-id (keyword)(opt)
   ;   :emptiable? (boolean)(constant)(opt)
   ;    Default: true